requirements.txt
fastapi
uvicorn
httptools
python-multipart
passlib
uvloop
bcrypt
python-jose
bleach
beautifulsoup4
